Переименовал компоненты, разбил на разные файлы
