+++ src/components/SentMessages/TableSentMessages/SentMessagesItem/SentMessagesItem.jsx
@@ -1,13 +1,14 @@
 /* Import libraries */
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import moment from 'moment';
-import ru from 'moment/locale/ru';
 
 /* Import reducers */
 import { updateMessageStatus } from '../../../../actionCreators/form';
 
-class SentMessagesItemContainer extends Component {
+/* Import components */
+import SentMessagesItemView from "./SentMessagesItemView";
+
+class SentMessagesItem extends Component {
   componentDidMount() {
     const { updateMessageStatus, id } = this.props;
 
@@ -29,31 +30,10 @@ class SentMessagesItemContainer extends Component {
   }
 
   render() {
-    return <SentMessagesItem {...this.props} />;
+    return <SentMessagesItemView {...this.props} />;
   }
 }
 
-const SentMessagesItem = ({ message, date, status }) => {
-  return (
-    <tr className="sent-messages-row">
-      <td className="sent-messages-data">
-        {date &&
-          moment(date)
-            .locale('ru')
-            .format('DD MMMM')}
-      </td>
-      <td className="sent-messages-data" title={message}>
-        {message}
-      </td>
-      <td className="sent-messages-data">
-        {status === -1 && <span className="color-success">Отправлено</span>}
-        {status < -1 && <span className="color-warning">Ошибка</span>}
-        {status > -1 && <span className="color-waiting">В процессе</span>}
-      </td>
-    </tr>
-  );
-};
-
 export default connect(() => ({}), { updateMessageStatus })(
-  SentMessagesItemContainer
+  SentMessagesItem
 );
